+++ tests/unit/MediawikiFactoryTest.php
@@ -38,6 +38,7 @@ class MediawikiFactoryTest extends \PHPUnit_Framework_TestCase {
 			array( 'Mediawiki\Api\Service\UserRightsChanger', 'newUserRightsChanger' ),
 			array( 'Mediawiki\Api\Service\UserCreator', 'newUserCreator' ),
 			array( 'Mediawiki\Api\Service\LogListGetter', 'newLogListGetter' ),
+			array( 'Mediawiki\Api\Service\FileUploader', 'newFileUploader' ),
 		);
 	}